refactor(web): extract pure helpers from useOnlineSkipBoGame
Move cloneGameStateFromView, applyOptimistic{Play,Discard}View,
collectDrawTransitions, inferOpponentTransition, etc. into
hooks/useOnlineSkipBoGame/helpers.ts. The hook file drops from
1123 to 898 LOC; the deeper split of the hook body into
connection/reconciler/optimistic-ui modules is left as follow-up.
